Update main.py
The trick of this project is not its complexity, but its usefulness and history.

I had an entrance exam at school, I wanted to write this exam for maximum score, but I could not understand how to solve one example, but I had already written something python, I did not understand how to solve this example, I wrote a code that solves this example by substitution and wrote a bot that I helped me to solve this example. I ran a local server and the bot and this bot helped me pass the paper with a maximum score.